Refactor into "ProcessStep"
