+++ source/dotnet/Services/IntegrationTests/Fixtures/WebApi/WebApiFactory.cs
@@ -13,10 +13,12 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Application;
-using Energinet.DataHub.Wholesale.Application.BatchActor;
+using Energinet.DataHub.Wholesale.Application.ProcessStep;
+using Energinet.DataHub.Wholesale.Application.ProcessStep.Model;
 using Energinet.DataHub.Wholesale.Application.SettlementReport;
 using Energinet.DataHub.Wholesale.Domain.Actor;
 using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 using Energinet.DataHub.Wholesale.Domain.SettlementReportAggregate;
 using Energinet.DataHub.Wholesale.WebApi;
 using Microsoft.AspNetCore.Authorization;
@@ -25,7 +27,7 @@ using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 
-namespace Energinet.DataHub.Wholesale.IntegrationTests.TestCommon.WebApi;
+namespace Energinet.DataHub.Wholesale.IntegrationTests.Fixtures.WebApi;
 
 /// <summary>
 /// When we execute the tests on build agents we use the builded output (assemblies).
@@ -73,10 +75,13 @@ public class WebApiFactory : WebApplicationFactory<Startup>
             services.AddScoped(
                 provider =>
                 {
-                    if (ActorApplicationServiceMock != null)
-                        return ActorApplicationServiceMock.Object;
+                    if (ProcessStepApplicationServiceMock != null)
+                        return ProcessStepApplicationServiceMock.Object;
 
-                    return new ActorApplicationService(provider.GetRequiredService<IActorRepository>());
+                    return new ProcessStepApplicationService(
+                        provider.GetRequiredService<IProcessStepResultRepository>(),
+                        provider.GetRequiredService<IProcessStepResultMapper>(),
+                        provider.GetRequiredService<IActorRepository>());
                 });
         });
     }
@@ -87,7 +92,7 @@ public class WebApiFactory : WebApplicationFactory<Startup>
     /// </summary>
     public Mock<ISettlementReportApplicationService>? SettlementReportApplicationServiceMock { get; set; }
 
-    public Mock<IActorApplicationService>? ActorApplicationServiceMock { get; set; }
+    public Mock<IProcessStepApplicationService>? ProcessStepApplicationServiceMock { get; set; }
 
     private sealed class AllowAnonymous : IAuthorizationHandler
     {
