+++ source/dotnet/Services/IntegrationTests/WebApi/BatchControllerTests.cs
@@ -15,6 +15,7 @@
 using System.Net;
 using System.Net.Http.Json;
 using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.IntegrationTests.Fixtures.WebApi;
 using Energinet.DataHub.Wholesale.IntegrationTests.TestCommon.Fixture.WebApi;
 using Energinet.DataHub.Wholesale.IntegrationTests.TestHelpers;
 using FluentAssertions;
