+++ source/dotnet/Services/UnitTests/Application/ProcessStep/Model/MarketRoleMapperTests.cs
@@ -12,16 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Application.ProcessStep.Model;
+using FluentAssertions;
+using Xunit;
+using Xunit.Categories;
+using MarketRole = Energinet.DataHub.Wholesale.Domain.Actor.MarketRole;
 
-namespace Energinet.DataHub.Wholesale.Application.ProcessResult;
+namespace Energinet.DataHub.Wholesale.Tests.Application.ProcessStep.Model;
 
-/// <summary>
-/// This class provides the ability to retrieve a calculated result for a given step for a batch.
-/// </summary>
-public interface IProcessStepResultApplicationService
+[UnitTest]
+public class MarketRoleMapperTests
 {
-    Task<ProcessStepResultDto> GetResultAsync(ProcessStepResultRequestDto processStepResultRequestDto);
-
-    Task<ProcessStepResultDto> GetResultAsync(ProcessStepResultRequestDtoV2 processStepResultRequestDtoV2);
+    [Theory]
+    [InlineData(Contracts.MarketRole.EnergySupplier, MarketRole.EnergySupplier)]
+    public void Map_Returns_CorrectMarketRoleType(
+        Contracts.MarketRole input,
+        MarketRole expected)
+    {
+        MarketRoleMapper.Map(input).Should().Be(expected);
+    }
 }
