+++ source/dotnet/Services/UnitTests/Application/ProcessStep/Model/ProcessResultPointTests.cs
@@ -12,20 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Application.ProcessResult;
+using Energinet.DataHub.Wholesale.Tests.TestHelpers;
+using Xunit;
+using Xunit.Categories;
 
-namespace Energinet.DataHub.Wholesale.Application.BatchActor;
+namespace Energinet.DataHub.Wholesale.Tests.Application.ProcessStep.Model;
 
-public static class MarketRoleTypeMapper
+[UnitTest]
+public sealed class ProcessResultPointTests
 {
-    public static Domain.Actor.MarketRole Map(MarketRole marketRole)
+    [Fact]
+    public async Task PropertyNamesAndTypesMatchContractWithCalculator()
     {
-        switch (marketRole)
-        {
-            case MarketRole.EnergySupplier:
-                return Domain.Actor.MarketRole.EnergySupplier;
-            default:
-                throw new ArgumentOutOfRangeException(nameof(marketRole), marketRole, null);
-        }
+        await using var stream = EmbeddedResources.GetStream("Sender.Infrastructure.Calculator.calculator-result.json");
+
+        await ContractComplianceTestHelper.VerifyTypeCompliesWithContractAsync<ProcessResultPoint>(stream);
     }
 }
