--- source/dotnet/Services/WebApi/Controllers/V2/BatchActorController.cs
@@ -1,43 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Energinet.DataHub.Wholesale.Application.BatchActor;
-using Energinet.DataHub.Wholesale.Contracts;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Energinet.DataHub.Wholesale.WebApi.Controllers.V2;
-
-[ApiController]
-[ApiVersion(Version)]
-[Route("v{version:apiVersion}/[controller]")]
-public class BatchActorController : ControllerBase
-{
-    private const string Version = "2.0";
-    private readonly IActorApplicationService _actorApplicationService;
-
-    public BatchActorController(IActorApplicationService actorApplicationService)
-    {
-        _actorApplicationService = actorApplicationService;
-    }
-
-    [AllowAnonymous] // TODO: Temporary hack to enable EDI integration while awaiting architects decision
-    [HttpPost]
-    [MapToApiVersion(Version)]
-    public async Task<IActionResult> GetAsync([FromBody] ProcessStepActorsRequest processStepActorsRequest)
-    {
-        var batchActorDtos = await _actorApplicationService.GetAsync(processStepActorsRequest).ConfigureAwait(false);
-        return Ok(batchActorDtos);
-    }
-}
